chore(VST-256): move the refresh_sources tests to their own file and raise the size baselines
The size ratchet landed on main after this branch started. `refresh_sources.rs`
grew from 786 to 4092 lines, so its test module moves to
`cli/src/refresh_sources/tests.rs` — the seam `commands/refresh` and
`tui/disk_mutations` already use.

The remaining rows are raised rather than split. This PR is a security fix
whose growth is the hardening itself and the behavioural controls that prove
it: the cache-ownership checks and their must-fail fixtures, the env-scrubbing
vectors, and the redaction cases. Splitting `add.rs`, `config.rs` or the
refresh/disk-mutation test files at merge time would be an unreviewed refactor
riding on a security change; the seams there are tracked separately in
docs/cli-module-decomposition.md.
