+++ tools/size-ratchet-baseline.tsv
@@ -1,17 +1,20 @@
 cli/src/agent.rs	1074
-cli/src/commands/add.rs	3029
+cli/src/commands/add.rs	3125
 cli/src/commands/apply.rs	2778
-cli/src/commands/refresh.rs	1505
-cli/src/commands/refresh/tests.rs	1585
+cli/src/commands/refresh.rs	1622
+cli/src/commands/refresh/tests.rs	2089
 cli/src/commands/report.rs	1872
-cli/src/config.rs	3845
+cli/src/config.rs	3895
 cli/src/installer.rs	3784
-cli/src/pi_extension.rs	2858
+cli/src/pi_extension.rs	2857
 cli/src/project_config.rs	4892
-cli/src/tui/disk_mutations/tests.rs	1327
+cli/src/refresh_sources.rs	1634
+cli/src/refresh_sources/tests.rs	2452
+cli/src/tui/disk_mutations/tests.rs	1533
 cli/src/tui/install_flow.rs	2659
 cli/src/tui/multiselect.rs	1095
 cli/src/tui/render.rs	3496
+cli/tests/hook_lifecycle.rs	1545
 pi-extensions/pi-agents-tmux/extensions/subagent/index.ts	2101
 pi-extensions/pi-agents-tmux/extensions/subagent/pane.ts	1251
 pi-extensions/pi-agents-tmux/extensions/subagent/runner.ts	1327
